requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-dotenv==0.19.1
openai==1.11.0
gunicorn==20.1.0
requests==2.28.1